requirements.txt
Flask
flask_cors
openai-whisper
torch
